Add name variable to User
User now has a String variable for usernames. Along with the expected getter and setter, there are two methods that, if given a name or ID, will return the other. Hopefully, this will allow future methods in other classes to not always require an entire User as an input for proper function