Change comments
The User constructor will need to be altered to generate the ID instead of requiring an ID be given to it. This ID will need to be generated to conform to a specific format and recorded to ensure duplication is prevented. Also, the commented methods were removed as they are unlikely to be used in the future